combine adder and calculator classes
